requirements.txt
# Requirements for the FastAPI backend
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.117.1
h11==0.16.0
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
python-dotenv==1.0.1
openai==1.109.1
pdfplumber==0.11.7
neo4j==5.28.2